ci: Automate integration tests (#450)
