--- integrationTests/.gitignore
@@ -1,6 +1,0 @@
-node_modules/
-hello/
-local.settings.json
-__pycache__/
-.python_packages/
-.serverless/
