--- integrationTests/configurations/node12-linux-premium/host.json
@@ -1,3 +1,0 @@
-{
-  "version": "2.0"
-}
