--- integrationTests/configurations/node12-linux-premium/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "node12-linux",
-  "version": "1.0.0",
-  "description": "Azure Functions sample for the Serverless framework",
-  "scripts": {
-    "test": "echo \"No tests yet...\"",
-    "start": "func host start"
-  },
-  "keywords": [
-    "azure",
-    "serverless"
-  ],
-  "dependencies": {},
-  "devDependencies": {
-    "serverless-azure-functions": "^1.0.0"
-  }
-}
