--- integrationTests/configurations/node12-linux-premium/serverless.yml
@@ -1,34 +1,0 @@
-service: nd12-lin-prem
-
-provider:
-  name: azure
-  prefix: tmp
-  stage: qa
-  region: West Europe
-  runtime: nodejs12
-  os: linux
-  apim: true
-  type: premium
-
-  environment:
-    VARIABLE_FOO: 'foo'
-
-plugins:
-  - serverless-azure-functions
-
-package:
-  exclude:
-    - local.settings.json
-    - .vscode/**
-
-functions:
-  hello:
-    handler: hello.sayHello
-    events:
-      - http: true
-        x-azure-settings:
-          methods:
-            - GET
-          authLevel: anonymous
-
-
