+++ integrationTests/configurations/node12-windows-premium/serverless.yml
@@ -4,7 +4,7 @@ provider:
   name: azure
   prefix: tmp
   stage: qa
-  region: West Europe
+  region: Central US
   runtime: nodejs12
   os: windows
   apim: true
@@ -26,9 +26,8 @@ functions:
     handler: hello.sayHello
     events:
       - http: true
-        x-azure-settings:
-          methods:
-            - GET
-          authLevel: anonymous
+        methods:
+          - GET
+        authLevel: anonymous
