+++ integrationTests/configurations/node12-windows-webpack/serverless.yml
@@ -26,9 +26,8 @@ functions:
     handler: hello.sayHello
     events:
       - http: true
-        x-azure-settings:
-          methods:
-            - GET
-          authLevel: anonymous
+        methods:
+          - GET
+        authLevel: anonymous
