--- integrationTests/models/parameters.ts
@@ -1,12 +1,0 @@
-import { Configurations } from "./configurations";
-import { InterpolateParameters } from "clvr";
-
-export interface DefaultParameters extends InterpolateParameters {
-  serviceName: string;
-  configName: string;
-  runtime: string;
-}
-
-export interface ConfigurationParameters extends Configurations {
-  [config: string]: DefaultParameters
-}
