--- integrationTests/tests/utils.ts
@@ -1,8 +1,0 @@
-import { readdirSync } from "fs";
-import { join } from "path"
-
-export function getDirectories(source = ".") {
-  return readdirSync(source, { withFileTypes: true })
-    .filter(dirent => dirent.isDirectory())
-    .map(dirent => join(source, dirent.name));
-}
